+++ src/index.ts
@@ -167,6 +167,7 @@ export type {
 } from './adt/IAdtProgram';
 export type {
   AdtFailureOrigin,
+  AdtNoFailure,
   IAdtError,
   IAdtFailure,
   IAdtResponse,
@@ -174,6 +175,7 @@ export type {
   IAdtSuccess,
   IResultStrategy,
 } from './adt/IAdtResponse';
+export { ADT_NO_FAILURE } from './adt/IAdtResponse';
 export type {
   ICreateScalarFunctionParams,
   IDeleteScalarFunctionParams,
@@ -386,7 +388,6 @@ export type {
 } from './runtime/IAtcRun';
 export type {
   ICrossTrace,
-  ICrossTraceDocuments,
   ICrossTraceResults,
   IListCrossTracesOptions,
 } from './runtime/ICrossTrace';
@@ -438,7 +439,6 @@ export type {
   IGenerateServiceBindingParams,
   IGetServiceBindingODataParams,
   IPublishODataV2Params,
-  IServiceBindingDocuments,
   IServiceBindingResults,
   ITransportCheckServiceBindingParams,
   IUnpublishODataV2Params,